Review Post Route Fixed

+++ db/schema.sql
@@ -10,7 +10,7 @@ product_name VARCHAR(100) NOT NULL,
 release_date TEXT NOT NULL,
 image TEXT DEFAULT 'https://dummyimage.com/400x400/6e6c6e/e9e9f5.png&text=No+Image',
 description TEXT NOT NULL,
-price INTEGER NOT NULL,
+price decimal (6,2) NOT NULL,
 category VARCHAR(50) NOT NULL,
 manufacturer TEXT NOT NULL
 );
